+++ tests/testthat/test-reduce.R
@@ -12,9 +12,14 @@ for (pattern in equivalent_reduction_patterns) {
             "reduce(x, '{pattern[[2]]}', '{reduction}')"
         ), {
             x <- create_seq_tensor(c(10, 20, 30, 40, 50))
+            if (inherits(x, "torch_tensor")) {
+                x <- x$to(dtype = torch::torch_float64())
+            }
             expect_no_error(out1 <- reduce(x, pattern[[1]], reduction))
             expect_no_error(out2 <- reduce(x, pattern[[2]], reduction))
             expect_identical(out1, out2)
         })
     }
 }
+
+# TODO more tests