+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## Additions
 
-- Add backend for `torch`'s `torch_tensor` objects
+- Add backend for `torch`'s `torch_tensor` objects for the three core functions `rearrange()`, `reduce()` and `"repeat"()`
 - `zeallot` is now a suggested dependency for users
 - `lintr` is now a suggested dependency for developers