+++ NEWS.md
@@ -8,6 +8,8 @@
 
 ## Changes
 
+- Minor updates in documentation and docker infrastructure in the sourcecode
+
 # einops 0.1.0
 
 - Initial release. The only python einops function implementations are `rearrange`, `reduce`, `repeat`, and `parse_shape`, and the only backend supported is for `base::array` objects. The first einops tutorial has also been mostly re-implemented as a vignette