+++ NEWS.md
@@ -2,10 +2,13 @@
 
 ## Additions
 
+- Add backend for `torch`'s `torch_tensor` objects for the three core functions `rearrange()`, `reduce()` and `"repeat"()`
 - `zeallot` is now a suggested dependency for users
 - `lintr` is now a suggested dependency for developers
 
 ## Changes
+
+- Minor updates in documentation and docker infrastructure in the sourcecode
 
 # einops 0.1.0